refactor: query handler (#977)
* refactor: query handler
