+++ packages/featureserver/src/route.spec.js
@@ -34,17 +34,12 @@ describe('Route module unit tests', () => {
       querySpy.calledOnce.should.equal(true);
       querySpy.firstCall.args.should.deepEqual([
         {
-          metadata: { maxRecordCount: 2000 },
+          params: { method: 'query' },
+          query: {},
+          url: '/FeatureServer/0/query',
         },
-        {
-          resultRecordCount: 2000,
-        },
-      ]);
-      responseHandlerSpy.calledOnce.should.equal(true);
-      responseHandlerSpy.firstCall.args.should.deepEqual([
         {},
-        { features: [] },
-        { resultRecordCount: 2000 },
+        {},
       ]);
     });
 
@@ -78,7 +73,7 @@ describe('Route module unit tests', () => {
             details: ['Fool bar'],
           },
         },
-        { resultRecordCount: 2000 },
+        {},
       ]);
     });
     afterEach(() => {
@@ -113,13 +108,11 @@ describe('Route module unit tests', () => {
       restInfoSpy.firstCall.args.should.deepEqual([
         {
           params: {},
-          query: { resultRecordCount: 2000 },
+          query: {},
           url: '/rest/info',
         },
         {},
-        {
-          metadata: { maxRecordCount: 2000 },
-        },
+        {},
       ]);
     });
 
@@ -145,13 +138,11 @@ describe('Route module unit tests', () => {
       restInfoSpy.firstCall.args.should.deepEqual([
         {
           params: {},
-          query: { resultRecordCount: 2000 },
+          query: {},
           url: '/rest/info',
         },
         {},
-        {
-          metadata: { maxRecordCount: 2000 },
-        },
+        {},
       ]);
 
       responseHandlerSpy.calledOnce.should.equal(true);
@@ -164,7 +155,7 @@ describe('Route module unit tests', () => {
             details: ['Fool bar'],
           },
         },
-        { resultRecordCount: 2000 },
+        {},
       ]);
     });
 
@@ -204,13 +195,11 @@ describe('Route module unit tests', () => {
       serverInfoSpy.firstCall.args.should.deepEqual([
         {
           params: {},
-          query: { resultRecordCount: 2000 },
+          query: {},
           url: '/rest/services/test/FeatureServer',
         },
         {},
-        {
-          metadata: { maxRecordCount: 2000 },
-        },
+        {},
       ]);
     });
 
@@ -236,13 +225,11 @@ describe('Route module unit tests', () => {
       serverInfoSpy.firstCall.args.should.deepEqual([
         {
           params: {},
-          query: { resultRecordCount: 2000 },
+          query: {},
           url: '/rest/services/test/FeatureServer',
         },
         {},
-        {
-          metadata: { maxRecordCount: 2000 },
-        },
+        {},
       ]);
     });
 
@@ -278,15 +265,11 @@ describe('Route module unit tests', () => {
       layersInfoSpy.firstCall.args.should.deepEqual([
         {
           params: {},
-          query: {
-            resultRecordCount: 2000,
-          },
+          query: {},
           url: '/rest/services/test/FeatureServer/layers',
         },
         {},
-        {
-          metadata: { maxRecordCount: 2000 },
-        },
+        {},
       ]);
     });
 
@@ -312,17 +295,11 @@ describe('Route module unit tests', () => {
       layersInfoSpy.firstCall.args.should.deepEqual([
         {
           params: {},
-          query: {
-            resultRecordCount: 2000,
-          },
+          query: {},
           url: '/rest/services/test/FeatureServer/layers',
         },
         {},
-        {
-          metadata: {
-            maxRecordCount: 2000,
-          },
-        },
+        {},
       ]);
 
       responseHandlerSpy.calledOnce.should.equal(true);
@@ -335,7 +312,7 @@ describe('Route module unit tests', () => {
             details: ['Fool bar'],
           },
         },
-        { resultRecordCount: 2000 },
+        {},
       ]);
     });
 
@@ -369,13 +346,11 @@ describe('Route module unit tests', () => {
       layerInfoSpy.firstCall.args.should.deepEqual([
         {
           params: {},
-          query: { resultRecordCount: 2000 },
+          query: {},
           url: '/rest/services/test/FeatureServer/0',
         },
         {},
-        {
-          metadata: { maxRecordCount: 2000 },
-        },
+        {},
       ]);
     });
 
@@ -401,13 +376,11 @@ describe('Route module unit tests', () => {
       layerInfoSpy.firstCall.args.should.deepEqual([
         {
           params: {},
-          query: { resultRecordCount: 2000 },
+          query: {},
           url: '/rest/services/test/FeatureServer/0',
         },
         {},
-        {
-          metadata: { maxRecordCount: 2000 },
-        },
+        {},
       ]);
     });
 
@@ -443,13 +416,11 @@ describe('Route module unit tests', () => {
       layerInfoSpy.firstCall.args.should.deepEqual([
         {
           params: {},
-          query: { resultRecordCount: 2000 },
+          query: {},
           url: '/rest/services/test/FeatureServer/0/info',
         },
         {},
-        {
-          metadata: { maxRecordCount: 2000 },
-        },
+        {},
       ]);
     });
 
@@ -480,12 +451,41 @@ describe('Route module unit tests', () => {
         {},
         {
           error: {
-            code: 404,
-            message: 'Not Found',
-            details: ['Not Found'],
+            code: 400,
+            message: 'Invalid URL',
+            details: ['Invalid URL'],
           },
         },
-        { resultRecordCount: 2000 },
+        {},
+      ]);
+    });
+
+    it('should reject unsupported layer method', () => {
+      const responseHandlerSpy = sinon.spy();
+      const route = proxyquire('./route', {
+        './response-handlers': { generalResponseHandler: responseHandlerSpy },
+      });
+
+      route(
+        {
+          params: {},
+          query: {},
+          url: '/rest/services/test/FeatureServer/0/goobar',
+        },
+        {},
+        {},
+      );
+      responseHandlerSpy.calledOnce.should.equal(true);
+      responseHandlerSpy.firstCall.args.should.deepEqual([
+        {},
+        {
+          error: {
+            code: 400,
+            message: 'Invalid URL',
+            details: ['Invalid URL'],
+          },
+        },
+        {},
       ]);
     });
   });
