+++ packages/output-geoservices/src/index.js
@@ -1,11 +1,5 @@
 const FeatureServer = require('@koopjs/featureserver');
-const {
-  restInfo,
-  serverInfo,
-  layerInfo,
-  layersInfo,
-  // query,
-} = require('@koopjs/featureserver');
+const { restInfo, serverInfo, layerInfo, layersInfo, query } = require('@koopjs/featureserver');
 const Logger = require('@koopjs/logger');
 let logger = new Logger();
 const ARCGIS_UNAUTHORIZED_MESSAGE = 'Item does not exist or is inaccessible.';
@@ -63,6 +57,11 @@ class GeoServices {
       path: '$namespace/rest/generateToken',
       methods: ['get', 'post'],
       handler: 'generateToken',
+    },
+    {
+      path: '$namespace/rest/services/$providerParams/FeatureServer/:layer/query',
+      methods: ['get', 'post'],
+      handler: 'queryHandler',
     },
     {
       path: '$namespace/rest/services/$providerParams/FeatureServer/:layer/:method',
@@ -223,6 +222,10 @@ class GeoServices {
 
   async layerInfoHandler(req, res) {
     this.#pullDataHandler(req, res, layerInfo);
+  }
+
+  async queryHandler(req, res) {
+    this.#pullDataHandler(req, res, query);
   }
 
   #buildTokensUrl(host, baseUrl) {
