+++ packages/output-geoservices/src/index.spec.js
@@ -9,6 +9,7 @@ jest.mock('@koopjs/featureserver', () => ({
   serverInfo: jest.fn(),
   layersInfo: jest.fn(),
   layerInfo: jest.fn(),
+  query: jest.fn(),
 }));
 
 const loggerMock = {
@@ -56,6 +57,11 @@ describe('Output Geoservices', () => {
           path: '$namespace/rest/generateToken',
           methods: ['get', 'post'],
           handler: 'generateToken',
+        },
+        {
+          path: '$namespace/rest/services/$providerParams/FeatureServer/:layer/query',
+          methods: ['get', 'post'],
+          handler: 'queryHandler',
         },
         {
           path: '$namespace/rest/services/$providerParams/FeatureServer/:layer/:method',
@@ -524,6 +530,17 @@ describe('Output Geoservices', () => {
           },
         },
       ]);
+    });
+  });
+
+  describe('queryHandler', () => {
+    test('should pull data and call handler', async () => {
+      const output = new OutputGeoServices(modelMock, { logger: loggerMock });
+      await output.queryHandler({ foo: 'bar' }, resMock);
+      expect(FeatureServer.query.mock.calls.length).toBe(1);
+      expect(FeatureServer.query.mock.calls[0]).toEqual([{ foo: 'bar' }, resMock, 'someData']);
+      expect(modelMock.pull.mock.calls.length).toBe(1);
+      expect(modelMock.pull.mock.calls[0][0]).toEqual({ foo: 'bar' });
     });
   });
